fix: preserve usage stats across pool resets + restore tab + split config
Usage stats:
- Add baseline to UsageStatsStore so historical totals survive account
  removal/replacement (previously getSummary read live entry.usage only)
- Restore UsageStats tab in Dashboard (lost in PR #211 layout refactor)
- UsageStats supports embedded prop for tab integration

Config split:
- Extract config-schema.ts (Zod schemas + types + constants, 72 LOC)
- Extract config-loader.ts (YAML load/merge/env override, 98 LOC)
- config.ts remains as barrel re-export + singleton state (115 LOC)
- All 40+ importers unchanged (barrel re-export)
